Add files via upload
This script compares a (nuceotide, currently) file against a BLAST db of interest. For each reference sequence in the blast DB, the best query hit (determined by best BIT score) is kept. Any sequences without a BLAST hit to the database are kept as well. Accordingly, the input query dataset is filtered for redundancy.