Update assembly2orf.sh
the previous version of assembly2orf linked to an old version of the pfamA file (the one I used during my PhD) because I couldn't get the new version to work on the server. While playing with Panther in 2017/2018 I managed to get the newer Pfam-A file working. It is found here:  /home/laura/data/external_data/Pfam/latestDownload_runFails/Pfam-A.hmm (unfortunately the directory has a stupid name but I dont want to change it). I edited lines 152, 193 (2x) and 194 to change the Pfam-A file used.